+++ google/cloud/google_cloud_cpp_common.bzl
@@ -1,5 +1,6 @@
 # DO NOT EDIT -- GENERATED BY CMake -- Change the CMakeLists.txt file if needed
 google_cloud_cpp_common_HDRS = [
+    "future.h",
     "iam_binding.h",
     "iam_bindings.h",
     "iam_policy.h",
@@ -10,6 +11,7 @@ google_cloud_cpp_common_HDRS = [
     "internal/future_base.h",
     "internal/future_fwd.h",
     "internal/future_impl.h",
+    "internal/future_then_impl.h",
     "internal/future_then_meta.h",
     "internal/future_generic.h",
     "internal/future_void.h",
