+++ google/cloud/internal/future_void.h
@@ -106,42 +106,7 @@ class future<void> final : private internal::future_base<void> {
   /// Implement `then()` if the result does not require unwrapping.
   template <typename F>
   typename internal::then_helper<F, void>::future_t then_impl(F&& functor,
-                                                              std::false_type) {
-    // g++-4.9 gets confused about the use of a protected type alias here, so
-    // create a non-protected one:
-    using local_state_type = shared_state_type;
-    // Some type aliases to make the rest of the code more readable.
-    using functor_result_t =
-        typename internal::then_helper<F, void>::functor_result_t;
-    using future_t = typename internal::then_helper<F, void>::future_t;
-
-    // The `shared_state_type` (aka `future_shared_state<void>`) is be written
-    // without any reference to the `future<void>` class, otherwise there would
-    // be cyclic dependencies between the two classes. We must adapt the
-    // provided functor, which takes a `future<void>` parameter to take a
-    // `shared_ptr<shared_state_type` parameter so it can be consumed by the
-    // underlying class. Because we need to support C++11, we use a local class
-    // instead of a lambda, as support for move+capture in lambdas is a C++14
-    // feature.
-    struct adapter {
-      explicit adapter(F&& func) : functor(func) {}
-
-      auto operator()(std::shared_ptr<local_state_type> state)
-          -> functor_result_t {
-        return functor(future<void>(std::move(state)));
-      }
-
-      F functor;
-    };
-
-    auto output_shared_state = shared_state_type::make_continuation(
-        shared_state_, adapter(std::forward<F>(functor)));
-
-    // Nothing throws after this point, and we have not changed the state if
-    // anything did throw.
-    shared_state_.reset();
-    return future_t(std::move(output_shared_state));
-  }
+                                                              std::false_type);
 
   /// Implement `then()` if the result requires unwrapping.
   // TODO(#1345) - implement this in a future PR.
